feat: create third thread (empty)
